fix: resolve CI linting issues for production release
- Fix errcheck issues: proper error handling for os.MkdirAll, fmt.Sscanf
- Fix gosec integer overflow warning in fleet target capacity
- Fix GitHub Actions gosec action (use direct install vs deprecated action)
- Fix import formatting issues
- Maintain Go Report Card A-grade compliance

Ready for clean v0.2.0 production release

+++ internal/aws/fleet.go
@@ -172,7 +172,7 @@ func (f *FleetManager) buildFleetRequest(req *FleetRequest, placementGroupName s
 			launchTemplateConfig,
 		},
 		TargetCapacitySpecification: &types.TargetCapacitySpecificationRequest{
-			TotalTargetCapacity:       aws.Int32(instanceCount),
+			TotalTargetCapacity:       aws.Int32(int32(instanceCount)),
 			DefaultTargetCapacityType: purchasingOption,
 		},
 		Type: types.FleetTypeInstant, // Synchronous launching for gang scheduling
